Added check to see if a symbolic link was present before making it for the 10m winds; added SERVER suffix to opendap notification email file name so that this file does not get overwritten when posting to multiple servers; made it possible to mix full path filenames and files assumed to be in the scenario directory when posting files to opendap.

+++ output/cpra_post.sh
@@ -95,7 +95,7 @@ if [[ -d $dirWind10m ]]; then
    logMessage "Corresponding 10m wind ensemble member was found."
    wind10mFound=yes
    for file in fort.72.nc fort.74.nc maxwvel.63.nc ; do
-      if [[ -e $dirWind10m/$file ]]; then
+      if [[ -e $dirWind10m/$file && ! -e ./wind10m.${file} ]]; then
          logMessage "$ENSTORM: $THIS: Found $dirWind10m/${file}."
          ln -s $dirWind10m/${file} ./wind10m.${file} 2>&1 | tee -a scenario.log >> $SYSLOG
          # update the run.properties file
